Defender Partially Working
Defender is partially working however it needs to be fixed, as when the robot does not see the goal it goes backwards with regular correction. It should simply go forwards with regular correction until it sees the goal, and once it does it can simply goal track. etc etc. (Should fix itself with line avoidance but we will see.)